==> stromaClinicalTrial/global/9-publish_and_push_to_synapse.Rmd <==
---
title: "Publish and push to Synapse"
output:
  html_document:
    df_print: paged
---


# Export crosstabs to txt file and push to synapse

```{r}
library(amlresistancenetworks)
library(dplyr)

library(reticulate)
syn = reticulate::import("synapseclient")
sync = syn$login()

path_to_data_folder <- "data"

for (msnset_path in list.files(path_to_data_folder, "msnset")) {
  load(file.path(path_to_data_folder, msnset_path))
  exprs(m) <- signif(exprs(m), 3)
  crosstab_path <- sub("msnset", "ptrc_ex17_crosstab", msnset_path)
  crosstab_path <- sub(".RData", ".txt", crosstab_path)
  write.table(exprs(m), file=file.path(path_to_data_folder, crosstab_path),
              quote=F, sep="\t")
}

synapse_folder <- "syn25572229"

for (crosstab_path in list.files(path_to_data_folder, "ptrc_ex17_crosstab.*corrected")) {
  synapseStore(file.path(path_to_data_folder, crosstab_path), synapse_folder)
}
```
