==> stromaClinicalTrial/global/10.1-merge_metadata_updated_and_push_to_synapse.Rmd <==
---
title: "Proteomic data processing"
author: "Camilo Posso"
date: "5/11/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
library(dplyr)
library(tibble)
```

## Data formats

Using the data available in the crosstabs and a second metadata file, we can create a Synapse table that is in 'long form' and has all the protein/phosphosite values together with their relevant metadata.

### Issues with SSL? Try disconnecting your VPN

```{r synapselogin}
source("../../util/synapseUtil.R")

syn <- synapseLogin()

tables.syn <- c("Global gene corrected Updated Metadata" = "syn25681742",
                "Global gene uncorrected Updated Metadata" = "syn25681732",
                "Global peptide corrected Updated Metadata" = "syn25681749",
                "Global peptide uncorrected Updated Metadata" = "syn25681747")

metadata.syn <- "syn26046812"

```


## Merge metadata and Global proteomics data


```{r pressure, echo=FALSE}
meta.tab <- readxl::read_xlsx(syn$get(metadata.syn)$path)
# meta.tab <- meta.tab %>%
#   mutate(Sample = as.character(Sample)) %>%
#   mutate(Sample = case_when(nchar(Sample) == 1 ~ paste0("0", Sample),
#                             TRUE ~ Sample))

tables <- list()

for (i in 1:length(tables.syn)){
  print(i)
  name <- names(tables.syn)[[i]]
  crosstab.syn <- tables.syn[[i]]
  mat <- read.csv(syn$get(crosstab.syn)$path,sep='\t',header=T)
  names(mat) <- sub("^X(.*)$","\\1",names(mat))
  mat <- add_column(mat, Accession = rownames(mat), .before = 1)
  
  mat <- mat%>%tidyr::pivot_longer(cols=c(2:ncol(mat)),names_to='Sample',
                      values_to='LogRatio')
  mat$Sample <- as.numeric(mat$Sample)
  mat <- left_join(mat,meta.tab, by = "Sample") %>% 
    subset(!is.na(LogRatio)) %>%
    arrange(Accession)
  tables[[name]] <- mat
}
```

Now we have all the data in long form.  The last thing that needs to be done is to store them both to a synapse table. This requires knowing the synapse project.

```{r store tables}
for (name in names(tables.syn)){
  print(name)
  mat <- tables[[name]]
  name <- paste("PTRC_ex17",name,sep = " ")
  print(name)
  synTableStore(mat,name,parentId='syn22128879')
}
```
