==> stromaClinicalTrial/global/12-GSEA_and_correlation_enrichment.Rmd <==
---
title: "Sunil Ex17 Global Correlation Enrichment"
author: "Camilo Posso"
date: "08/17/2021"
output: 
  html_document:
    code_folding: hide
    toc: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r Setup libraries, include=FALSE}
library(amlresistancenetworks)
library(kableExtra)
library(dplyr)
library(readr)
library(readxl)
library(limma)
source("../../Util/synapseUtil.R")

# corrected for loading mass and plexID, long time and subtle
g.gene.corrected <- querySynapseTable("syn26047126")

prot.dat <- g.gene.corrected

## We have the metadata selected from the long form table
summary <- g.gene.corrected[1:40, c("Sample", "BeatAML Patient ID", "Plex", 
                                    "Loading Mass", "Description", "Period", "Treatment")]
rownames(summary) <- summary$Sample

## Organizing time points into the right order (not alphanumeric)
time.points <- c("P2 Pre-Study", "P2 Cycle 1 Day 1", "P2 Cycle 1 Day 28", "P2 Cycle 2 Day 28", "P2 Cycle 3 Day 28", 
                 "P1B Cycle 1 Day 1", "P1B Cycle 1 Day 28", "P1B Cycle 2 Day 28", "P1B Cycle 3 Day 28", "P1B Cycle 6 Day 28",
                 "P1B Cycle 9 Day 28", "P1B Conc Cycle 9 Day 28") %>%
  factor(.)


summary$Description <- factor(summary$Description, levels = time.points)
summary <- summary[sort(summary$Sample), ]
summary$`BeatAML Patient ID` <- factor(summary$`BeatAML Patient ID`, levels = unique(summary$`BeatAML Patient ID`))
summary$Treatment <- make.names(summary$Treatment)
```

## Introduction

We access synapse to load the data in long form. We use data which has been corrected for both
Loading mass effect, as well as plex effect. Below we see a metadata table for this experiment.

```{r metadata, echo=FALSE}
show <- summary
kbl(show) %>%
  kable_paper("hover", full_width = T)

```

```{r prepare groups and matrix, include=FALSE}
##select the samples of interest
healthy <- subset(summary, `BeatAML Patient ID` == 'Healthy Donor Stroma')

pre.treatment <- subset(summary, Period == "Pre-Treatment")

## Earliest sample, per patient, which has been treated with GILT
treated.gilt <- subset(summary, Treatment == "GILT") %>%
  filter(`BeatAML Patient ID` != "5029") %>%
  group_by(`BeatAML Patient ID`) %>%
  slice(1)

## Earliest sample, per patient, which has been treated with GILT + DEC
treated.gilt.and.dec <- subset(summary, Treatment == "GILT...DEC") %>%
  group_by(`BeatAML Patient ID`) %>%
  slice(1)

##then we spread the proteomics data into a matrix
colnames(prot.dat)[[1]] <- "Gene"
prot.mat <- prot.dat%>%
  select(LogRatio,Sample,Gene)%>%
  tidyr::pivot_wider(values_from='LogRatio',names_from='Sample',
                     values_fn=list(LogRatio=mean))%>%
  tibble::column_to_rownames('Gene')
```

In this report we collect both rank based GSEA for each stated comparison, as well the results of
Jason McDermott's correlation enrichment. 

Correlation enrichment aims to identify pathways whose *genes* are highly correlated.
Within a group of samples, this approach can help identify pathways whose genes show a high degree of concordance.
For a pathway \(\mathcal{P}\), we compute the correlation between all 
*pairs of genes* belonging to \(\mathcal{P}\), and define the `Correlation Enrichment` 
as the average correlation over all pairs. Values closer to 1 indicate that the genes 
in \(\mathcal{P}\) are tightly correlated; this is the value we plot for each pathway below. 
The p-value measures whether the difference in between the `Correlation Enrichment` and the 
background average correlation among genes is significant.

We split the report according to which database we use to perform the analysis.

## Using KEGG


### Healthy vs Pre-Treatment

#### GSEA

Below we have a GSEA plot computed using the KEGG pathway
database. We're comparing the healthy donor samples to Pre-Treatment samples. 
Red indicates increased activity in Pre-Treatment samples, while blue shows increased activity
in healthy samples. 

```{r, include=TRUE, fig.align="center", fig.cap=c("Healthy vs Pre-Treatment")}
knitr::include_graphics("sig-includedkegg Healthy vs Pre-Treatment-gsea-plot.png")
```

#### Correlation Enrichment

Below we show the correlation enrichment in Healthy samples on the left, 
as well as Pre-Treatment samples on the right. Pathways with higher enrichment score indicate more 
concordance among the genes in that pathway. Note that color does not have any
meaning in correlation enrichment.


```{r, echo=FALSE,out.width="49%", out.height = "49%", fig.cap="Left: Healthy - Right: Pre-Treatment",fig.show='hold',fig.align='center'}
knitr::include_graphics(c("sig-included-Kegg Healthy-correlation-enrichment-plot.png",
                          "sig-included-Kegg Pre-Treatment-correlation-enrichment-plot.png"))
``` 

$$\\[0.2in]$$

### Healthy vs Gilteritinib

#### GSEA

Below we have a GSEA plot computed using the KEGG pathway
database. We compare healthy samples to Gilteritinib treated samples.
To avoid over-representing any patient, we choose the earliest sample per patient for the Gilteritinib group.
Red indicates increased activity in Gilteritinib samples, while blue shows increased activity
in healthy samples.

```{r, include=TRUE, fig.align="center", fig.cap=c("Healthy vs Gilteritinib")}
knitr::include_graphics("sig-includedkegg Healthy vs Gilteritinib-gsea-plot.png")
```

#### Correlation Enrichment

Below we show the correlation enrichment in Healthy samples on the left, 
as well as Gilteritinib samples on the right. Pathways with higher enrichment score indicate more 
concordance among the genes in that pathway. Note that color does not have any
meaning in correlation enrichment.

```{r, echo=FALSE,out.width="49%", out.height = "49%", fig.cap="Left: Healthy - Right: Gilteritinib",fig.show='hold',fig.align='center'}
knitr::include_graphics(c("sig-included-Kegg Healthy-correlation-enrichment-plot.png",
                          "sig-included-Kegg Gilteritinib-correlation-enrichment-plot.png"))
``` 

$$\\[0.2in]$$

### Healthy vs Gilteritinib + Decitabine

#### GSEA

Below we have a GSEA plot computed using the KEGG pathway
database. Next, we compare healthy samples to Gilteritinib + Decitabine treated samples.
To avoid over-representing any patient, we choose the earliest sample per patient for the Gilteritinib Decitabine group.
Red indicates increased activity in Gilteritinib Decitabine samples, while blue shows increased activity
in healthy samples.

```{r, include=TRUE, fig.align="center", fig.cap=c("Healthy vs Gilteritinib Decitabine")}
knitr::include_graphics("sig-includedkegg Healthy vs Gilteritinib Decibatine-gsea-plot.png")
```

#### Correlation Enrichment

Below we show the correlation enrichment in Healthy samples on the left, 
as well as Gilteritinib + Decitabine samples on the right. Pathways with higher enrichment score indicate more 
concordance among the genes in that pathway. Note that color does not have any
meaning in correlation enrichment.


```{r, echo=FALSE,out.width="49%", out.height = "49%", fig.cap="Left: Healthy - Right: Gilteritinib Decitabine",fig.show='hold',fig.align='center'}
knitr::include_graphics(c("sig-included-Kegg Healthy-correlation-enrichment-plot.png",
                          "sig-included-Kegg Gilteritinib Decitabine-correlation-enrichment-plot.png"))
``` 

$$\\[0.2in]$$

### Pre-Treatment vs Gilteritinib

#### GSEA

Below we have a GSEA plot computed using the KEGG pathway
database. Next, we compare Pre-Treatment samples to Gilteritinib treated samples.
To avoid over-representing any patient, we choose the earliest sample per patient for the Gilteritinib group.
Red indicates increased activity in Gilteritinib samples, while blue shows increased activity
in Pre-Treatment.

```{r, include=TRUE, fig.align="center", fig.cap=c("Pre-Treatment vs Gilteritinib")}
knitr::include_graphics("sig-includedkegg Pre-Treatment vs Gilteritinib-gsea-plot.png")
```

#### Correlation Enrichment

Below we show the correlation enrichment in Pre-Treatment samples on the left, 
as well as Gilteritinib samples on the right. Pathways with higher enrichment score indicate more 
concordance among the genes in that pathway. Note that color does not have any
meaning in correlation enrichment.

```{r, echo=FALSE,out.width="49%", out.height = "49%", fig.cap="Left: Pre-Treatment - Right: Gilteritinib",fig.show='hold',fig.align='center'}
knitr::include_graphics(c("sig-included-Kegg Pre-Treatment-correlation-enrichment-plot.png",
                          "sig-included-Kegg Gilteritinib-correlation-enrichment-plot.png"))
``` 

$$\\[.2in]$$

### Gilteritinib vs Gilteritinib + Decitabine

#### GSEA

Below we have a GSEA plot computed using the KEGG pathway
database. Next, we compare Gilteritinib samples to Gilteritinib + Decitabine treated samples.
To avoid over-representing any patient, we choose the earliest sample per patient in both groups.
The Gilteritinib group in this comparison comprises only patients with a corresponding Gilteritinib + Decitabine sample.
Red indicates increased activity in Gilteritinib + Decitabine samples, while blue shows increased activity
in Gilteritinib samples.

```{r, include=TRUE, fig.align="center", fig.cap=c("Gilteritinib vs Gilteritinib + Decitabine")}
knitr::include_graphics("sig-includedkegg Gilteritinib vs Gilteritinib Decitabine-gsea-plot.png")
```

#### Correlation Enrichment

Below we show the correlation enrichment in Gilteritinib samples on the left, 
as well as Gilteritinib + Decitabine samples on the right. Pathways with higher enrichment score indicate more 
concordance among the genes in that pathway. Note that color does not have any
meaning in correlation enrichment.

```{r, echo=FALSE,out.width="49%", out.height = "49%", fig.cap="Left: Gilteritinib - Right: Gilteritinib + Decitabine",fig.show='hold',fig.align='center'}
knitr::include_graphics(c("sig-included-Kegg Gilteritinib Restricted-correlation-enrichment-plot.png",
                          "sig-included-Kegg Gilteritinib Decitabine-correlation-enrichment-plot.png"))
``` 

$$\\[.2in]$$


## Using REACTOME


### Healthy vs Pre-Treatment

#### GSEA

Below we have a GSEA plot computed using the REACTOME pathway
database. We're comparing the healthy donor samples to Pre-Treatment samples. 
Red indicates increased activity in Pre-Treatment samples, while blue shows increased activity
in healthy samples. 

```{r, include=TRUE, fig.align="center", fig.cap=c("Healthy vs Pre-Treatment")}
knitr::include_graphics("sig-includedreactome Healthy vs Pre-Treatment-gsea-plot.png")
```

#### Correlation Enrichment

Below we show the correlation enrichment in Healthy samples on the left, 
as well as Pre-Treatment samples on the right. Pathways with higher enrichment score indicate more 
concordance among the genes in that pathway. Note that color does not have any
meaning in correlation enrichment.


```{r, echo=FALSE,out.width="49%", out.height = "49%", fig.cap="Left: Healthy - Right: Pre-Treatment",fig.show='hold',fig.align='center'}
knitr::include_graphics(c("sig-included-REACTOME Healthy-correlation-enrichment-plot.png",
                          "sig-included-REACTOME Pre-Treatment-correlation-enrichment-plot.png"))
``` 

$$\\[0.2in]$$

### Healthy vs Gilteritinib

#### GSEA

Below we have a GSEA plot computed using the REACTOME pathway
database. We compare healthy samples to Gilteritinib treated samples.
To avoid over-representing any patient, we choose the earliest sample per patient for the Gilteritinib group.
Red indicates increased activity in Gilteritinib samples, while blue shows increased activity
in healthy samples.

```{r, include=TRUE, fig.align="center", fig.cap=c("Healthy vs Gilteritinib")}
knitr::include_graphics("sig-includedreactome Healthy vs Gilteritinib-gsea-plot.png")
```

#### Correlation Enrichment

Below we show the correlation enrichment in Healthy samples on the left, 
as well as Gilteritinib samples on the right. Pathways with higher enrichment score indicate more 
concordance among the genes in that pathway. Note that color does not have any
meaning in correlation enrichment.

```{r, echo=FALSE,out.width="49%", out.height = "49%", fig.cap="Left: Healthy - Right: Gilteritinib",fig.show='hold',fig.align='center'}
knitr::include_graphics(c("sig-included-REACTOME Healthy-correlation-enrichment-plot.png",
                          "sig-included-REACTOME Gilteritinib-correlation-enrichment-plot.png"))
``` 

$$\\[0.2in]$$

### Healthy vs Gilteritinib + Decitabine

#### GSEA

Below we have a GSEA plot computed using the REACTOME pathway
database. Next, we compare healthy samples to Gilteritinib + Decitabine treated samples.
To avoid over-representing any patient, we choose the earliest sample per patient for the Gilteritinib Decitabine group.
Red indicates increased activity in Gilteritinib Decitabine samples, while blue shows increased activity
in healthy samples.

```{r, include=TRUE, fig.align="center", fig.cap=c("Healthy vs Gilteritinib Decitabine")}
knitr::include_graphics("sig-includedreactome Healthy vs Gilteritinib Decibatine-gsea-plot.png")
```

#### Correlation Enrichment

Below we show the correlation enrichment in Healthy samples on the left, 
as well as Gilteritinib + Decitabine samples on the right. Pathways with higher enrichment score indicate more 
concordance among the genes in that pathway. Note that color does not have any
meaning in correlation enrichment.


```{r, echo=FALSE,out.width="49%", out.height = "49%", fig.cap="Left: Healthy - Right: Gilteritinib Decitabine",fig.show='hold',fig.align='center'}
knitr::include_graphics(c("sig-included-REACTOME Healthy-correlation-enrichment-plot.png",
                          "sig-included-REACTOME Gilteritinib Decitabine-correlation-enrichment-plot.png"))
``` 

$$\\[0.2in]$$

### Pre-Treatment vs Gilteritinib

#### GSEA

Below we have a GSEA plot computed using the REACTOME pathway
database. Next, we compare Pre-Treatment samples to Gilteritinib treated samples.
To avoid over-representing any patient, we choose the earliest sample per patient for the Gilteritinib group.
Red indicates increased activity in Gilteritinib samples, while blue shows increased activity
in Pre-Treatment.

```{r, include=TRUE, fig.align="center", fig.cap=c("Pre-Treatment vs Gilteritinib")}
knitr::include_graphics("sig-includedreactome Pre-Treatment vs Gilteritinib-gsea-plot.png")
```

#### Correlation Enrichment

Below we show the correlation enrichment in Pre-Treatment samples on the left, 
as well as Gilteritinib samples on the right. Pathways with higher enrichment score indicate more 
concordance among the genes in that pathway. Note that color does not have any
meaning in correlation enrichment.

```{r, echo=FALSE,out.width="49%", out.height = "49%", fig.cap="Left: Pre-Treatment - Right: Gilteritinib",fig.show='hold',fig.align='center'}
knitr::include_graphics(c("sig-included-REACTOME Pre-Treatment-correlation-enrichment-plot.png",
                          "sig-included-REACTOME Gilteritinib-correlation-enrichment-plot.png"))
``` 

$$\\[.2in]$$

### Gilteritinib vs Gilteritinib + Decitabine

#### GSEA

Below we have a GSEA plot computed using the REACTOME pathway
database. Next, we compare Gilteritinib samples to Gilteritinib + Decitabine treated samples.
To avoid over-representing any patient, we choose the earliest sample per patient in both groups.
The Gilteritinib group in this comparison comprises only patients with a corresponding Gilteritinib + Decitabine sample.
Red indicates increased activity in Gilteritinib + Decitabine samples, while blue shows increased activity
in Gilteritinib samples.

```{r, include=TRUE, fig.align="center", fig.cap=c("Gilteritinib vs Gilteritinib + Decitabine")}
knitr::include_graphics("sig-includedreactome Gilteritinib vs Gilteritinib Decitabine-gsea-plot.png")
```

#### Correlation Enrichment

Below we show the correlation enrichment in Gilteritinib samples on the left, 
as well as Gilteritinib + Decitabine samples on the right. Pathways with higher enrichment score indicate more 
concordance among the genes in that pathway. Note that color does not have any
meaning in correlation enrichment.

```{r, echo=FALSE,out.width="49%", out.height = "49%", fig.cap="Left: Gilteritinib - Right: Gilteritinib + Decitabine",fig.show='hold',fig.align='center'}
knitr::include_graphics(c("sig-included-REACTOME Gilteritinib Restricted-correlation-enrichment-plot.png",
                          "sig-included-REACTOME Gilteritinib Decitabine-correlation-enrichment-plot.png"))
``` 

$$\\[.2in]$$





[comment]: # (Below is the code to produce all the plots seen in this markdown)

```{r limma for GSEA, eval=FALSE, include=FALSE,, include=FALSE}
## LogFC computed as pre-treatment - healthy
prot.diffex.pre <- limmaTwoFactorDEAnalysis(prot.mat, healthy$Sample, pre.treatment$Sample) %>%
  rename(value = "logFC") %>%
  mutate(Gene = rownames(.))

## LogFC computed as treated GILT - healthy.
prot.diffex.gilt <- limmaTwoFactorDEAnalysis(prot.mat, healthy$Sample, treated.gilt$Sample) %>%
  rename(value = "logFC") %>%
  mutate(Gene = rownames(.))

## LogFC computed as treated GILT+DEC - healthy
prot.diffex.gilt.dec <- limmaTwoFactorDEAnalysis(prot.mat, healthy$Sample, treated.gilt.and.dec$Sample) %>%
  rename(value = "logFC") %>%
  mutate(Gene = rownames(.))


#### Paired comparisons for Pre vs GILT and GILT vs GILT + DEC ####
## Pre vs GILT. logFC computed as GILT - PRE
samples.compared <- rbind(pre.treatment, treated.gilt)
ID <- factor(samples.compared$`BeatAML Patient ID`, levels = unique(samples.compared$`BeatAML Patient ID`))
Treatment <- factor(samples.compared$Treatment, levels = c("PRE", "GILT"))
design <- model.matrix(~ID+Treatment)
fit <- lmFit(prot.mat[, samples.compared$Sample], design)
fit <- eBayes(fit)
prot.paired.gilt.vs.pre <- topTable(fit, coef = "TreatmentGILT", number = Inf) %>%
  mutate(featureID = rownames(.),
         Gene = rownames(.)) %>%
  rename(value = logFC) %>%
  arrange(P.Value)

### To make sure the moderated t-statistics resemble those from a t.test
# t.test(as.numeric(prot.mat["RAB5B", treated.gilt$Sample]),
#        as.numeric(prot.mat["RAB5B", pre.treatment$Sample]),
#        paired = T)

## GILT vs GILT+DEC. logFC computed as GILT...DEC - GILT
## Choosing only GILT samples with corresponding GILT+DEC 
treated.gilt.restricted <- subset(summary, Treatment == "GILT") %>%
  filter(`BeatAML Patient ID` %in% treated.gilt.and.dec$`BeatAML Patient ID`)
samples.compared <- rbind(treated.gilt.restricted, treated.gilt.and.dec)
ID <- factor(samples.compared$`BeatAML Patient ID`, levels = unique(samples.compared$`BeatAML Patient ID`))
Treatment <- factor(samples.compared$Treatment, levels = c("GILT", "GILT...DEC"))
design <- model.matrix(~ID+Treatment)
fit <- lmFit(prot.mat[, samples.compared$Sample], design)
fit <- eBayes(fit)
prot.paired.giltdec.vs.gilt <- topTable(fit, coef = "TreatmentGILT...DEC", number = Inf) %>%
  mutate(featureID = rownames(.),
         Gene = rownames(.)) %>%
  rename(value = logFC) %>%
  arrange(P.Value)
```

```{r KEGG GSEA, eval=FALSE, include=FALSE}
library(tidyr)
library(leapr)
data(ncipid)
data("krbpaths")

idx.kegg <- grepl("^KEGG_", krbpaths$names)
names.kegg <- krbpaths$names[idx.kegg]
names.kegg <- sub("KEGG_", "", names.kegg)
names.kegg <- gsub("_", " ", names.kegg)
names.kegg <- sapply(names.kegg, function(y) paste(strwrap(y, 45), 
                                          collapse = "\n"), 
                     USE.NAMES = FALSE)
desc.kegg <- krbpaths$desc[idx.kegg]
sizes.kegg <- krbpaths$sizes[idx.kegg]
Max <- max(sizes.kegg)
matrix.kegg <- krbpaths$matrix[idx.kegg, 1:Max]
keggpaths <- list(names = names.kegg,
                 desc = desc.kegg,
                 sizes = sizes.kegg,
                 matrix = matrix.kegg)

term.2.gene <- as.data.frame(keggpaths$matrix) %>%
  mutate(term = keggpaths$names) %>%
  pivot_longer(!term, names_to = "Column", values_to = "gene") %>%
  filter(!(gene == "null")) %>%
  select(term, gene)

term.2.name <- data.frame(term = keggpaths$names, name = keggpaths$names)
  
plotGSEA(prot.diffex.pre, term.2.gene = term.2.gene, 
         term.2.name = term.2.name, prefix = "kegg Healthy vs Pre-Treatment")

plotGSEA(prot.diffex.gilt, term.2.gene = term.2.gene, 
         term.2.name = term.2.name, prefix = "kegg Healthy vs Gilteritinib")

plotGSEA(prot.diffex.gilt.dec, term.2.gene = term.2.gene, 
         term.2.name = term.2.name, prefix = "kegg Healthy vs Gilteritinib Decibatine")

plotGSEA(prot.paired.gilt.vs.pre, term.2.gene = term.2.gene, 
         term.2.name = term.2.name, prefix = "kegg Pre-Treatment vs Gilteritinib")

plotGSEA(prot.paired.giltdec.vs.gilt, term.2.gene = term.2.gene, 
         term.2.name = term.2.name, prefix = "kegg Gilteritinib vs Gilteritinib Decitabine")

```

```{r REACTOME GSEA, eval=FALSE, include=FALSE}
library(tidyr)
library(leapr)
idx.reactome <- grepl("^REACTOME_", krbpaths$names)
names.reactome <- krbpaths$names[idx.reactome]
names.reactome <- sub("REACTOME_", "", names.reactome)
names.reactome <- gsub("_", " ", names.reactome)
names.reactome <- sapply(names.reactome, function(y) paste(strwrap(y, 45), 
                                          collapse = "\n"), 
                         USE.NAMES = FALSE)
desc.reactome <- krbpaths$desc[idx.reactome]
sizes.reactome <- krbpaths$sizes[idx.reactome]
Max <- max(sizes.reactome)
matrix.reactome <- krbpaths$matrix[idx.reactome, 1:Max]
reactomepaths <- list(names = names.reactome,
                 desc = desc.reactome,
                 sizes = sizes.reactome,
                 matrix = matrix.reactome)

term.2.gene <- as.data.frame(reactomepaths$matrix) %>%
  mutate(term = reactomepaths$names) %>%
  pivot_longer(!term, names_to = "Column", values_to = "gene") %>%
  filter(!(gene == "null")) %>%
  filter(!(gene == "")) %>%
  select(term, gene)

term.2.name <- data.frame(term = reactomepaths$names, name = reactomepaths$names)

plotGSEA(prot.diffex.pre, term.2.gene = term.2.gene, 
         term.2.name = term.2.name, prefix = "reactome Healthy vs Pre-Treatment")

plotGSEA(prot.diffex.gilt, term.2.gene = term.2.gene, 
         term.2.name = term.2.name, prefix = "reactome Healthy vs Gilteritinib")

plotGSEA(prot.diffex.gilt.dec, term.2.gene = term.2.gene, 
         term.2.name = term.2.name, prefix = "reactome Healthy vs Gilteritinib Decibatine")

plotGSEA(prot.paired.gilt.vs.pre, term.2.gene = term.2.gene, 
         term.2.name = term.2.name, prefix = "reactome Pre-Treatment vs Gilteritinib")

plotGSEA(prot.paired.giltdec.vs.gilt, term.2.gene = term.2.gene, 
         term.2.name = term.2.name, prefix = "reactome Gilteritinib vs Gilteritinib Decitabine")

```

```{r enrichment correlation KEGG, eval=FALSE, include=FALSE}
library(tidyr)
library(leapr)
data(ncipid)
data("krbpaths")

idx.kegg <- grepl("^KEGG_", krbpaths$names)
names.kegg <- krbpaths$names[idx.kegg]
names.kegg <- sub("KEGG_", "", names.kegg)
names.kegg <- gsub("_", " ", names.kegg)
names.kegg <- sapply(names.kegg, function(y) paste(strwrap(y, 45), 
                                          collapse = "\n"), 
                     USE.NAMES = FALSE)
desc.kegg <- krbpaths$desc[idx.kegg]
sizes.kegg <- krbpaths$sizes[idx.kegg]
Max <- max(sizes.kegg)
matrix.kegg <- krbpaths$matrix[idx.kegg, 1:Max]
keggpaths <- list(names = names.kegg,
                 desc = desc.kegg,
                 sizes = sizes.kegg,
                 matrix = matrix.kegg)

enrichment.correlation.healthy <- plotCorrelationEnrichment(exprs = prot.mat[, healthy$Sample], 
                                                            prefix = "Kegg Healthy", pathway.plot.size = 4, 
                                                            geneset = keggpaths, corr.cutoff = 0.4,
                                                            clean.names = FALSE, height = 12)

enrichment.correlation.pre <- plotCorrelationEnrichment(exprs = prot.mat[,pre.treatment$Sample], 
                                                        prefix = "Kegg Pre-Treatment", pathway.plot.size = 4,
                                                        geneset = keggpaths, corr.cutoff = 0.4,
                                                        clean.names = FALSE, height = 12)

enrichment.correlation.treated.gilt <- plotCorrelationEnrichment(exprs = prot.mat[,treated.gilt$Sample], 
                                                                 prefix = "Kegg Gilteritinib", pathway.plot.size = 4,
                                                                 geneset = keggpaths, corr.cutoff = 0.4,
                                                                 clean.names = FALSE, height = 12)

enrichment.correlation.treated.giltdec <- plotCorrelationEnrichment(exprs = prot.mat[, treated.gilt.and.dec$Sample], 
                                                                    prefix = "Kegg Gilteritinib Decitabine", pathway.plot.size = 4,
                                                                    geneset = keggpaths, corr.cutoff = 0.4,
                                                                    clean.names = FALSE, height = 12)

enrichment.correlation.treated.gilt.restricted <- plotCorrelationEnrichment(exprs = prot.mat[, treated.gilt.restricted$Sample], 
                                                                            prefix = "Kegg Gilteritinib Restricted", 
                                                                            pathway.plot.size = 4,
                                                                            geneset = keggpaths, corr.cutoff = 0.4,
                                                                            clean.names = FALSE, height = 12)
```

```{r enrichment correlation REACTOME, eval=FALSE, include=FALSE}
library(tidyr)
library(leapr)
idx.reactome <- grepl("^REACTOME_", krbpaths$names)
names.reactome <- krbpaths$names[idx.reactome]
names.reactome <- sub("REACTOME_", "", names.reactome)
names.reactome <- gsub("_", " ", names.reactome)
names.reactome <- sapply(names.reactome, function(y) paste(strwrap(y, 45), 
                                          collapse = "\n"), 
                         USE.NAMES = FALSE)
desc.reactome <- krbpaths$desc[idx.reactome]
sizes.reactome <- krbpaths$sizes[idx.reactome]
Max <- max(sizes.reactome)
matrix.reactome <- krbpaths$matrix[idx.reactome, 1:Max]
reactomepaths <- list(names = names.reactome,
                 desc = desc.reactome,
                 sizes = sizes.reactome,
                 matrix = matrix.reactome)

enrichment.correlation.healthy <- plotCorrelationEnrichment(exprs = prot.mat[, healthy$Sample], 
                                                            prefix = "REACTOME Healthy", pathway.plot.size = 4, 
                                                            geneset = reactomepaths, corr.cutoff = 0.4,
                                                            clean.names = FALSE, height = 12)

enrichment.correlation.pre <- plotCorrelationEnrichment(exprs = prot.mat[,pre.treatment$Sample], 
                                                        prefix = "REACTOME Pre-Treatment", pathway.plot.size = 4,
                                                        geneset = reactomepaths, corr.cutoff = 0.4,
                                                        clean.names = FALSE, height = 12)

enrichment.correlation.treated.gilt <- plotCorrelationEnrichment(exprs = prot.mat[,treated.gilt$Sample], 
                                                                 prefix = "REACTOME Gilteritinib", pathway.plot.size = 4,
                                                                 geneset = reactomepaths, corr.cutoff = 0.4,
                                                                 clean.names = FALSE, height = 12)

enrichment.correlation.treated.giltdec <- plotCorrelationEnrichment(exprs = prot.mat[, treated.gilt.and.dec$Sample], 
                                                                    prefix = "REACTOME Gilteritinib Decitabine", 
                                                                    pathway.plot.size = 4,
                                                                    geneset = reactomepaths, corr.cutoff = 0.4,
                                                                    clean.names = FALSE, height = 12)

enrichment.correlation.treated.gilt.restricted <- plotCorrelationEnrichment(exprs = prot.mat[, treated.gilt.restricted$Sample], 
                                                                            prefix = "REACTOME Gilteritinib Restricted", 
                                                                            pathway.plot.size = 4,
                                                                            geneset = reactomepaths, corr.cutoff = 0.4,
                                                                            clean.names = FALSE, height = 12)
```
