==> stromaClinicalTrial/global/14-GSEA_patient_heatmaps.Rmd <==
---
title: "Sunil Ex17 Pathway heatmaps"
author: "Camilo Posso"
date: "08/17/2021"
output: 
  html_document:
    code_folding: hide
    toc: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

### Metadata

```{r libraries and setup, include=FALSE}
library(amlresistancenetworks)
library(kableExtra)
library(dplyr)
library(readr)
library(readxl)
library(limma)
source("../../Util/synapseUtil.R")

# corrected for loading mass and plexID, long time and subtle
g.gene.corrected <- querySynapseTable("syn26047126")

prot.dat <- g.gene.corrected

## We have the metadata selected from the long form table
summary <- g.gene.corrected[1:40, c("Sample", "BeatAML Patient ID", "Plex", 
                                    "Loading Mass", "Description", "Period", "Treatment")]
rownames(summary) <- summary$Sample

## Organizing time points into the right order (not alphanumeric)
time.points <- c("P2 Pre-Study", "P2 Cycle 1 Day 1", "P2 Cycle 1 Day 28", "P2 Cycle 2 Day 28", "P2 Cycle 3 Day 28", 
                 "P1B Cycle 1 Day 1", "P1B Cycle 1 Day 28", "P1B Cycle 2 Day 28", "P1B Cycle 3 Day 28", "P1B Cycle 6 Day 28",
                 "P1B Cycle 9 Day 28", "P1B Conc Cycle 9 Day 28") %>%
  factor(.)


summary$Description <- factor(summary$Description, levels = time.points)
summary <- summary[sort(summary$Sample), ]
summary$`BeatAML Patient ID` <- factor(summary$`BeatAML Patient ID`, levels = unique(summary$`BeatAML Patient ID`))
summary$Treatment <- make.names(summary$Treatment)
```

Below we show only the metadata for the 7 patients.

```{r metadata, echo=FALSE}
show <- summary
kbl(show) %>%
  kable_paper("hover", full_width = T)

```

$$\\[0.5in]$$ 

### GSEA Heatmaps Gilteritinib vs Pre-treatment
 
Six of the seven patients have both a Pre-treatment and Gilteritinib sample. From these 6 patients,
We have a total of 6 Pre-Treatment samples (one per patient), each with a single corresponding Gilteritinib sample, except
patient 5145, who has 2 Gilteritinib samples.

We use these 6 patients to run GSEA to compare the Pre-Treatment group to the Gilteritinib group on a per patient basis.
After collecting these enrichment results, we re-adjust 
the p-values using "BH" (Benjamini Hochberg) correction. We select only the top 50 most enriched
pathways and select those with an adjusted p-value below 0.05. 

In these plots, red indicates increased activity in the Gilteritinib, while blue shows increased
activity in the Pre-Treatment group. Moreover, the asterisk indicates an adjusted p-value below 0.05 for 
that particular patient + pathway.

We use the net enrichment score for each pathway + patient combination in the heatmap.

```{r out.width="49%", out.height = "49%", fig.show='hold', fig.align='center', fig.cap=c("LEFT: KEGG ---- RIGHT: REACTOME")}
knitr::include_graphics("Making Plots/Pathway heatmap Gilteritinib vs Pre-Treatment - KEGG.png")
knitr::include_graphics("Making Plots/Pathway heatmap Gilteritinib vs Pre-Treatment - REACTOME.png")
```

### GSEA Heatmaps Gilteritinib + Decitabine vs Gilteritinib

Only three of the seven patients have a Gilteritinib + Decitabine sample, namely "5029", "5104" and "5174". 
Patients "5029" and "5104" both have one sample for each of the Gilteritinib and Gilteritinib + Decitabine groups, 
while "5104" has one Gilteritinib and 4 Gilteritinib + Decitabine sample.

We use these 3 patients to run GSEA to compare the Late period 
to the Early period for each patient. After collecting these enrichment results, we re-adjust 
the p-values using "BH" (Benjamini Hochberg) correction. We select only the top 50 most enriched
pathways and select those with an adjusted p-value below 0.05. 

In these plots, red indicates increased activity in the Gilteritinib + Decitabine group, while blue shows increased
activity in the Gilteritinib group. Moreover, the asterisk indicates an adjusted p-value below 0.05 for 
that particular patient + pathway.

We use the net enrichment score for each pathway + patient combination in the heatmap.

```{r out.width="49%", out.height = "49%", fig.show='hold', fig.align='center', fig.cap=c("LEFT: KEGG ---- RIGHT: REACTOME")}
knitr::include_graphics("Making Plots/Pathway heatmap Gilteritinib + Decitabine vs Gilteritinib - KEGG.png")
knitr::include_graphics("Making Plots/Pathway heatmap Gilteritinib + Decitabine vs Gilteritinib - REACTOME.png")
```
